requirements.txt
pandas
numpy<2
tensorflow==2.16.2
keras==3.12.0
matplotlib
mlflow
scikit-learn
customtkinter
Pillow